+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Unreleased
+
 Added
   * AZ-768: Add `monitoring_function_app_service_plan_name`
+
+Fixed
+  * AZ-588: Add parameters to configure FAME storage network rules (required by table queries Terraform resources)
 
 # v6.0.1 - 2022-05-18